requirements.txt
google-generativeai==0.8.5
flask
geopy
reportlab==4.2.2
dotenv
flask-cors